+++ lib/constant/constant.dart
@@ -1,3 +1,0 @@
-export 'src/area.dart';
-export 'src/constant.dart';
-export 'src/way.dart';