+++ lib/extended/scroll_view/scroll_view.dart
@@ -4,9 +4,6 @@ import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_waya/flutter_waya.dart';
 
-export 'extended_scroll_view.dart';
-export 'list_view.dart';
-export 'sliver/sliver.dart';
 
 /// 可刷新的滚动组件
 /// 嵌套 sliver 家族组件