+++ lib/utils/utils.dart
@@ -1,12 +1,6 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter_waya/flutter_waya.dart';
 
-export 'src/ali_oss.dart';
-export 'src/des/des.dart';
-export 'src/des/des_engine.dart';
-export 'src/des/permuted.dart';
-export 'src/event.dart';
-export 'src/media_query.dart';
 
 void logDebug(dynamic msg) => debugPrint(msg.toString());